Pull the geocoding functionality into a separate module
